+++ tests/testthat/test-ard_emmeans_contrast.R
@@ -1,27 +1,26 @@
 skip_if_not(is_pkg_installed(c("emmeans", "survey", "lme4")))
 
-test_that("ard_emmeans_mean_difference() works", {
+test_that("ard_emmeans_contrast() works", {
   withr::local_options(width = 250)
 
-  expect_error(
-    ard_emmeans_mean_difference <-
-      ard_emmeans_mean_difference(
+  expect_silent(
+    ard_emmeans_contrast <-
+      ard_emmeans_contrast(
         data = mtcars,
         formula = vs ~ am + mpg,
         method = "glm",
         method.args = list(family = binomial),
         response_type = "dichotomous"
-      ),
-    NA
+      )
   )
-  expect_snapshot(ard_emmeans_mean_difference |> print(columns = "all"))
+  expect_snapshot(ard_emmeans_contrast |> print(columns = "all"))
 
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "method")[1],
+    cards::get_ard_statistics(ard_emmeans_contrast, stat_name %in% "method")[1],
     list(method = "Least-squares adjusted mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "mean.difference.estimate") |>
+    cards::get_ard_statistics(ard_emmeans_contrast, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     glm(vs ~ am + mpg, data = mtcars, family = binomial) |>
@@ -31,35 +30,23 @@ test_that("ard_emmeans_mean_difference() works", {
       getElement("estimate")
   )
 
-  expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference, stat_name %in% "mean.estimate") |>
-      unlist() |>
-      unname(),
-    glm(vs ~ am + mpg, data = mtcars, family = binomial) |>
-      emmeans::emmeans(specs = ~am, regrid = "response") |>
-      summary(infer = TRUE) |>
-      getElement("prob")
-  )
-
-
-  expect_error(
-    ard_emmeans_mean_difference_lme4 <-
-      ard_emmeans_mean_difference(
+  expect_silent(
+    ard_emmeans_contrast_lme4 <-
+      ard_emmeans_contrast(
         data = mtcars,
         formula = vs ~ am + (1 | cyl),
         method = "glmer",
         method.args = list(family = binomial),
         package = "lme4",
         response_type = "dichotomous"
-      ),
-    NA
+      )
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "method")[1],
+    cards::get_ard_statistics(ard_emmeans_contrast_lme4, stat_name %in% "method")[1],
     list(method = "Least-squares mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "mean.difference.estimate") |>
+    cards::get_ard_statistics(ard_emmeans_contrast_lme4, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     lme4::glmer(vs ~ am + (1 | cyl), data = mtcars, family = binomial) |>
@@ -67,37 +54,26 @@ test_that("ard_emmeans_mean_difference() works", {
       emmeans::contrast(method = "pairwise") |>
       summary(infer = TRUE) |>
       getElement("estimate")
-  )
-
-  expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_lme4, stat_name %in% "mean.estimate") |>
-      unlist() |>
-      unname(),
-    lme4::glmer(vs ~ am + (1 | cyl), data = mtcars, family = binomial) |>
-      emmeans::emmeans(specs = ~am, regrid = "response") |>
-      summary(infer = TRUE) |>
-      getElement("prob")
   )
 
   #styler: off
-  expect_error({
+  expect_silent({
     data(api, package = "survey")
-    ard_emmeans_mean_difference_svy <-
+    ard_emmeans_contrast_svy <-
       survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc) |>
-      ard_emmeans_mean_difference(
+      ard_emmeans_contrast(
         formula = api00 ~ sch.wide,
         method = "svyglm",
         package = "survey"
-      )},
-    NA
+      )}
   )
   # styler: on
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "method")[1],
+    cards::get_ard_statistics(ard_emmeans_contrast_svy, stat_name %in% "method")[1],
     list(method = "Least-squares mean difference")
   )
   expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "mean.difference.estimate") |>
+    cards::get_ard_statistics(ard_emmeans_contrast_svy, stat_name %in% "estimate") |>
       unlist() |>
       unname(),
     survey::svyglm(api00 ~ sch.wide, design = survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)) |>
@@ -106,20 +82,11 @@ test_that("ard_emmeans_mean_difference() works", {
       summary(infer = TRUE) |>
       getElement("estimate")
   )
-  expect_equal(
-    cards::get_ard_statistics(ard_emmeans_mean_difference_svy, stat_name %in% "mean.estimate") |>
-      unlist() |>
-      unname(),
-    survey::svyglm(api00 ~ sch.wide, design = survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)) |>
-      emmeans::emmeans(specs = ~sch.wide, regrid = "response") |>
-      summary(infer = TRUE) |>
-      getElement("emmean")
-  )
 })
 
-test_that("ard_emmeans_mean_difference() follows ard structure", {
+test_that("ard_emmeans_contrast() follows ard structure", {
   expect_silent(
-    ard_emmeans_mean_difference(
+    ard_emmeans_contrast(
       data = mtcars,
       formula = vs ~ am + mpg,
       method = "glm",
@@ -130,11 +97,11 @@ test_that("ard_emmeans_mean_difference() follows ard structure", {
   )
 })
 
-test_that("ard_emmeans_mean_difference() errors are returned correctly", {
+test_that("ard_emmeans_contrast() errors are returned correctly", {
   withr::local_options(width = 250)
 
   expect_silent(
-    ard <- ard_emmeans_mean_difference(
+    ard <- ard_emmeans_contrast(
       data = mtcars,
       formula = vs ~ am + mpg,
       method = "glm",